Feat: 몹 과 나무 배치 완료 및 맵 구성 완료
몹 스포너와 몹이 나무를 통과하지 못하게 직접 나무를 오브젝트로 배치

맵 가장자리에 벽을 세워 플레이어가 나가지 못하게 막음

몹이 일렬로 최대한 못오게끔 수정

+++ Assets/1. Script/SO/Monster/MonsterSO.cs
@@ -8,7 +8,7 @@ public class MonsterSO : ScriptableObject
     public int _maxHp;
     public int _attackDamage;
     public float _attackSpeed =1;
-    public int _moveSpeed;
+    public float _moveSpeed;
     public GameObject _enemyModelPrefab; 
     public float _attackRange;
 }
